files/data/scripts/requirements.txt
# Pip requirements
PySide6-Essentials>=6.7.2
PyOpenGL>=3.1.7